Update README.md
A museum in Italy wants to predict the number of subscribers for the following year using customer data. To achieve this, we will use customer data (age, gender, average spending, type of subscription...) and build three types of models using supervised learning (AI) algorithms. We will estimate profit and draw a profit curve based on the above predictions.